Feat(clean): removed redundant variables + change testing (a bit) + moved the hash logic completely to the backend
